Some fixes for the model library.
Default attributes are now used when undefined values for a given
attributes are seen when loading new model data. For example a model
with an attributes named "interfaceId" that defaults to 2 will now
correctly set this value to 2 after loading data from the server with
data = {interfaceId: undefined}. Undefined values are undefined and
null, or NaN for all decimal / integer values.
Previously the attrbute would have been NaN for number atrributes, or
"undefined", "null", etc for strings. Not good...

some other formatting and minor bugfixes