silence "Empty string passed to getElementById()." in tk_graph.js:578
The webdeveloper log is spammed with warnings as we use an empty name as
key in to check if a function with this name (doesn't) exist. An empty
string will not only print a console warning through, it will also
trigger the non-existence, so we silence the warning by making this
behavior explicit.