prevent applyBinding to the same element multiple times
SBPM is a single-page application, so the page isn't reloaded in its
entirely, but just certain parts replaced. Especially the div#main is
the element most used to change its childs, but given that #main remains
the same knockout is right in seeing that we bind multiple ViewModels
against it, so we ensure now that we bind against a child, which is
properly destroyed on page change and knockout is happy.