add a script to create various screenshots as simple tests
Testing websites is hard, but it is also hard to change anything if all
testing is only anecdotal… This script isn't the solution to this
problem, but it can at least help in finding changes caused by
refactoring or library updates. Note that this isn't testing any
usecase or anything interactive, so there is still much opportunity for
breakage not covered by this. This is just a very humble start…