update jquery to v1.11.2
This 'breaks' a bunch of its plugins, so those are updated as well to
their latest versions. They mostly break by not being loaded through,
so every instance I found was corrected, but many are still potentially
missing. Not much point in blowing this commit up even further
especially as this isn't strictly related to jquery to start with.