reorganize files belonging to the frontend
In the previously split up directories it can be quite confusing which
file is actually belonging to an external dependencies and which one is
created by us for our use, so instead of using this split in css, js and
image files we drop everything belonging to an external dependencies in
its own directory to make this obvious and therefore simpler to upgrade.

In the process a few old (or 'legacy') files are dropped entirely as
they have no current use and in case we would ever need them, we have
them in the git history.