drop jquery-ui in favor of bootstrap datepicker
jQuery UI is a big framework and we just used the datepicker from it.
And not even completely as the CSS rules couldn't be enabled as they
conflicted with other rules we have. The new datepicker on the other
hand is specialized, does only datepicking and is styled for bootstrap.