fix moment's deprecation warning for "unformated" inputs
moment used to accept any string as input and would give it if it
couldn't detect a format to JavaScripts Date. Upstream deems this to
fragile through and recommends changing this - or at least make it
explicit. We use it to parse a string formatted by our datepicker, so we
know which format to expect and hence can hand it to moment in this case
to resolve the issue for good.