give up on using require.js for jquery associates
tk_graph is needing all sorts of external dependencies, but it isn't
using require.js and it is very non-trivial to teach it to, which means
to have it operate stable jquery and all its plugins (especially pubsub)
need to be loaded once and before tk_graph, which works most of the
time, but becomes very fragile with additions. The next commit e.g.
re-introduces feedback, which 'just' contains an ajax call: the result
being that a reload on a subject-view causes no pubsubs to register
anymore, which cripples the entire frontend (without any error message
of course, just that nothing works anymore…).