independent scrolling for graph and settings on process
Scrolling the whole page is no fun, so with some rework we can confine
the graph and its settings to a div each, which with some exact
positioning are made so that we don't scroll the whole page but each div
individual if they are larger than the display port itself.

This is kinda what we had before, just in a cleaned up way compatible
with all the other changes.