improve "unsaved changes" behaviour
The loadView of Router does the unsaved check, but some routes called
the loadView of App directly… (not an ideal API, but lets keep it for
now). We now consistently call the Router.loadView so that every page
will show the message correctly.

We also keep track of the last path we were on now, so that in the
process page we can keep the unsaved state even while navigating between
process and subject view so that we don't need to display the message
here anymore.