remove unused expandListOfProcesses()
In commit 486bd479d1853ccfdc27c96f1c6370002c8df9c2 the element which is
clicked in via this method cascade was replaced with the breadcrumb.
It was used to expand the "List of processes" menu option which used to
have all exisiting processes as submenu entries. Given the amount of
processes we have nowadays this seems rather unlikely to come back soon…