fix chardin.js introduction display
chardin uses 'data-intro' instead of 'data-chardin-intro' in newer
versions (which are also the fields used by intro.js), so a sed over all
files and everything is mostly fine… expect that limiting the intro to
a specific element doesn't work with our styling. The solution here is to
use always 'body' as root element which looks better, too. We just have
to make sure that our always shown elements (e.g. menu) are excluded
from the intro in the general case.