SBPM-1005: try harder to detect an already created graph
The code tries to detect if a graph is already created, but the check is
always true, so that the special case for modeling by restriction if the
graph isn't already created is never called. Via a deeper inspection we
can figure out if the process comes from the database or not and hence
base our decision of created graph or not on that (as well).